refactor: String -> Enum 변환 로직 enum 클래스 내부로 이동
