+++ src/main/java/com/web/stard/domain/report/domain/entity/Report.java
@@ -2,7 +2,7 @@ package com.web.stard.domain.report.domain.entity;
 
 import com.web.stard.domain.member.domain.entity.Member;
 import com.web.stard.domain.post.domain.enums.PostType;
-import com.web.stard.domain.reply.domain.enums.ReportReason;
+import com.web.stard.domain.report.domain.enums.ReportReason;
 import com.web.stard.global.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
