+++ src/main/java/com/web/stard/global/config/security/SecurityConfig.java
@@ -57,6 +57,7 @@ public class SecurityConfig {
                         .requestMatchers(HttpMethod.GET, "/faqs/**").permitAll()
                         .requestMatchers(HttpMethod.GET, "/faqs-and-qnas/**").permitAll()
                         .requestMatchers(HttpMethod.GET, "/qnas/**").permitAll()
+                        .requestMatchers(HttpMethod.GET, "/replies/**").permitAll()
                         .anyRequest().authenticated() // 다른 모든 요청은 인증 필요
                 )
                 .sessionManagement(session -> {
